refactor: 구조 변경 및 랜덤 인터페이스를 생성자로 받도록 수정
