+++ src/main/kotlin/carRacing/controller/main.kt
@@ -1,6 +1,7 @@
 package carRacing
 
-import carRacing.controller.Game
+import carRacing.model.Game
+import carRacing.util.Random
 import carRacing.view.InputView
 import carRacing.view.ResultView
 
@@ -9,7 +10,8 @@ fun main() {
     val numMovement: Int = InputView.numMovement()
 
     ResultView.printTitle()
-    val game = Game(carNameList, numMovement)
+    val randomGenerator: Random = Random.RandomGenerator()
+    val game = Game(carNameList, numMovement, randomGenerator)
     game.run()
     ResultView.printResultRecord(carNameList, game.recordList)
     ResultView.printWinner(game.winner())
